+++ src/Phrase.cpp
@@ -4,7 +4,7 @@
 
 #include "Phrase.hpp"
 
-Phrase::Phrase(std::uint32_t w, std::string p) : weight(w), phrase(p) {}
+Phrase::Phrase(std::uint64_t w, std::string p) : weight(w), phrase(p) {}
 
 bool Phrase::operator<(Phrase const &rhs) const {
   return this->phrase < rhs.phrase;
